modify

+++ include/ColumnTable.h
@@ -10,19 +10,19 @@ class ColumnTable
 {
 private:
 
-    typedef std::vector<std::pair<std::string, size_t> > vector_t;
+    typedef std::vector<std::pair<std::string, uint64_t> > vector_t;
 
 public:
 
     ColumnTable(std::string tableName);
     virtual ~ColumnTable();
-    void insertOneDGroupKey(void* insertPtr, std::string columnName, size_t type);
-    std::pair<void*, size_t> getOneDGroupKey(std::string columnName);
-    std::vector<uint64_t> updateOneDGroupKey(void* updatePtr, std::string columnName, size_t type);
+    void insertOneDGroupKey(void* insertPtr, std::string columnName, uint64_t type);
+    std::pair<void*, uint64_t> getOneDGroupKey(std::string columnName);
+    std::vector<uint64_t> updateOneDGroupKey(void* updatePtr, std::string columnName, uint64_t type);
     std::string getTableName();
     void print();
 
-    inline size_t getPos(std::string columnName)//precise search, if not exist, return -1
+    inline uint64_t getPos(std::string columnName)//precise search, if not exist, return -1
     {
         for(vector_t::iterator itr = m_indexVector.begin(); itr != m_indexVector.end();
             itr++)
